// packages/common/src/message-types.ts
export const APP_HELLO = "app-hello"

export type AppHelloArgs = {
  userSessionId: string
  instanceId: string
  appId: string
}
// ...
export type SailAppStateArgs = AppRegistration[]

/**
 * These two messages carry FDC3 Communication Protocol messages.
 */
export const FDC3_APP_EVENT = "fdc3-app-event" // from the app to the server
export const FDC3_DA_EVENT = "fdc3-da-event" // from the server to the app